Updated search.tsx to improve usability of the app
As a first time user of Issure, it took me a while to understand that after running yarn dev and opening the localhost:3000 page, I had to enter text in the search box. As a beginner I think this description change in my PR will help other beginners as well.

Thanks for creating an awesome simple app - it was a breeze to open this up in vscode and setup the dev environment.